# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "EvalCell": {
        "description": "One cell of the eval matrix: a case's result on a version.\n\n``model`` is the model the result was produced under (the matrix's model\ndimension), or ``None`` when the recording run carried no model tag.\n\n``detail`` is the scorer's optional explanation for the verdict. It is\ndistinct from a trace's ``error``, which identifies a turn that did not\ncomplete.\n\n``plumbing_ok`` means the case ran to completion but no grader judged it (the\nfake-model tier, ADR-0055). It is a distinct status rather than a pass or a\nfail because it is neither: the fake answers from a canned script, so its cell\ncarries no comparative information and must never read as a green promotion\ngate.",
        "properties": {
          "case_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Count"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "scorer": {
            "anyOf": [
              {
                "enum": [
                  "grader",
                  "trajectory"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scorer"
          },
          "status": {
            "enum": [
              "pass",
              "fail",
              "plumbing_ok",
              "missing"
            ],
            "title": "Status",
            "type": "string"
          },
          "stream_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream Id"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version",
          "status"
        ],
        "title": "EvalCell",
        "type": "object"
      },
      "EvalMatrix": {
        "description": "The eval matrix grid: rows = cases, columns = versions (most recent first).\n\n``models`` and ``model_summaries`` add the model dimension: the distinct\nmodels observed across the fetched traces, and a pass-rate + cost rollup per\nmodel for BYO-model comparison. ``model_version_summaries`` slices that same\nrollup per ``(version, model)`` so a caller can scope completion to a single\ntriggered sha rather than the blended window (#814). They are additive; the\nversion grid is unchanged.",
        "properties": {
          "cases": {
            "items": {
              "type": "string"
            },
            "title": "Cases",
            "type": "array"
          },
          "model_summaries": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/EvalModelSummary"
            },
            "title": "Model Summaries",
            "type": "array"
          },
          "model_version_summaries": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/EvalModelVersionSummary"
            },
            "title": "Model Version Summaries",
            "type": "array"
          },
          "models": {
            "default": [],
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Models",
            "type": "array"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/EvalMatrixRow"
            },
            "title": "Rows",
            "type": "array"
          },
          "suite": {
            "title": "Suite",
            "type": "string"
          },
          "versions": {
            "items": {
              "type": "string"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "required": [
          "suite",
          "versions",
          "cases",
          "rows"
        ],
        "title": "EvalMatrix",
        "type": "object"
      },
      "EvalMatrixRow": {
        "description": "One row of the eval matrix: a case across every version column.",
        "properties": {
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "cells": {
            "items": {
              "$ref": "#/components/schemas/EvalCell"
            },
            "title": "Cells",
            "type": "array"
          }
        },
        "required": [
          "case_id",
          "cells"
        ],
        "title": "EvalMatrixRow",
        "type": "object"
      },
      "EvalModelSummary": {
        "description": "A per-model rollup across the suite: pass-rate and total cost.\n\nThe model dimension of the matrix (issue #255): the same suite run across\nmodels is sliceable here into ``passed/total`` pass-rate and summed\n``cost_usd`` per model, so BYO-model work can compare which models a use case\ntolerates and at what cost. ``cost_usd`` is ``None`` when no case under this\nmodel reported a cost (e.g. the fake-model path), rather than a misleading 0.\n\n``plumbing`` counts the rows that ran but were never graded (ADR-0055). They\nare excluded from ``passed``/``total``: counted as passes the fake model reads\n100% and counted as fails it reads 0%, both fabricated. The count keeps those\nrows visible instead of silently dropping them, so a model whose only rows are\nplumbing still appears with ``total == 0``.\n\n``completed`` counts the graded rows (within ``total``) whose turn actually\nreached a verdict, as opposed to a graded FAIL that never completed at all\n(a classified failure, a turn that ended in the wrong terminal status, or a\ntransport/runner exception -- see ``EvalCaseResult.error`` in the worker).\n``total`` alone cannot tell a real 0% (every case completed and the grader\nsaid no) apart from a model that never produced one completed turn (issue\n#622, #526 AC4): a model whose id does not resolve, or whose runner boots but\nnever answers, drives every case through the SAME classified-failure path a\nreal model's bad answer never touches. A sweep row with ``total > 0`` and\n``completed == 0`` is that distinct outcome, not a real (if unlucky) 0%.",
        "properties": {
          "completed": {
            "default": 0,
            "title": "Completed",
            "type": "integer"
          },
          "cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Usd"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "passed": {
            "title": "Passed",
            "type": "integer"
          },
          "plumbing": {
            "default": 0,
            "title": "Plumbing",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "passed",
          "total"
        ],
        "title": "EvalModelSummary",
        "type": "object"
      },
      "EvalModelVersionSummary": {
        "description": "A per-(version, model) rollup: the graded aggregates scoped to a single\nversion column, not rolled across the whole shown window.\n\n``EvalModelSummary`` sums ``completed`` over EVERY in-window version for a\nmodel. That blend can mask a triggered sha that lands all-incomplete (the\nmodel boots but never completes a turn on the new code) when a prior in-window\nsha completed cases for that same model: the blended ``completed`` stays ``> 0``\nfrom the old sha, so the \"never completed\" outcome the sweep must fail on\n(ADR-0068, #622) is hidden and a blended pass-rate is reported as a real\ncomparison (issue #814). This per-version breakdown exposes the\n``(version, model)`` dimension so a caller -- the CLI ``--model`` sweep, which\nknows the sha it just triggered -- can scope ``completed``/never-completed to\nthat one sha instead of the window.\n\nFields mirror the graded subset of ``EvalModelSummary`` (``cost_usd`` is not\nsliced per version, since the sweep does not compare cost per sha). It is\nadditive and defaulted the way ``completed``/``plumbing`` already are: a caller\nthat predates the field reads an empty list and degrades to the blended\nreading rather than misreporting.",
        "properties": {
          "completed": {
            "default": 0,
            "title": "Completed",
            "type": "integer"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "passed": {
            "title": "Passed",
            "type": "integer"
          },
          "plumbing": {
            "default": 0,
            "title": "Plumbing",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version",
          "passed",
          "total"
        ],
        "title": "EvalModelVersionSummary",
        "type": "object"
      },
      "EvalReport": {
        "description": "An eval run's rollup, reported so the API can post the PR check.\n\nByte-identical across both lanes before this promotion, so it carries no\nsemantic decisions.",
        "properties": {
          "passed_count": {
            "title": "Passed Count",
            "type": "integer"
          },
          "repo_full_name": {
            "title": "Repo Full Name",
            "type": "string"
          },
          "sha": {
            "title": "Sha",
            "type": "string"
          },
          "target_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Url"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "repo_full_name",
          "sha",
          "passed_count",
          "total"
        ],
        "title": "EvalReport",
        "type": "object"
      },
      "EvalReportResult": {
        "description": "The committed GitHub commit-status state for a reported eval run.",
        "properties": {
          "sha": {
            "title": "Sha",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "state",
          "sha"
        ],
        "title": "EvalReportResult",
        "type": "object"
      },
      "EvalTriggerRequest": {
        "description": "Ask for an on-demand platform eval run for an agent (issue #10).\n\nEnqueues the same EvalJob the git-push fan-out uses, minus the\npush-only gate. With no version_id the agent's active dev deployment is\nevaluated; suite falls back to Settings.eval_default_suite when omitted.",
        "properties": {
          "agent_id": {
            "format": "uuid",
            "title": "Agent Id",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "suite": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suite"
          },
          "target_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Url"
          },
          "version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version Id"
          }
        },
        "required": [
          "agent_id"
        ],
        "title": "EvalTriggerRequest",
        "type": "object"
      },
      "EvalTriggerResult": {
        "description": "The enqueued eval job's stream id plus the resolved job identity.",
        "properties": {
          "agent_id": {
            "format": "uuid",
            "title": "Agent Id",
            "type": "string"
          },
          "bundle_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bundle Ref"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "sha": {
            "title": "Sha",
            "type": "string"
          },
          "stream_id": {
            "title": "Stream Id",
            "type": "string"
          },
          "suite": {
            "title": "Suite",
            "type": "string"
          },
          "version_id": {
            "format": "uuid",
            "title": "Version Id",
            "type": "string"
          }
        },
        "required": [
          "stream_id",
          "agent_id",
          "version_id",
          "sha",
          "suite",
          "bundle_ref"
        ],
        "title": "EvalTriggerResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Curie API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/evals/matrix": {
      "get": {
        "operationId": "eval_matrix_evals_matrix_get",
        "parameters": [
          {
            "in": "query",
            "name": "suite",
            "required": true,
            "schema": {
              "title": "Suite",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "versions",
            "required": false,
            "schema": {
              "default": 5,
              "title": "Versions",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "stream_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stream Id"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvalMatrix"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Eval Matrix",
        "tags": [
          "evals"
        ]
      }
    },
    "/evals/report": {
      "post": {
        "operationId": "report_eval_evals_report_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalReport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvalReportResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Report Eval",
        "tags": [
          "evals"
        ]
      }
    },
    "/evals/trigger": {
      "post": {
        "operationId": "trigger_eval_evals_trigger_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalTriggerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvalTriggerResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trigger Eval",
        "tags": [
          "evals"
        ]
      }
    },
...
  }
}